Fix issue where ball could "escape" table
Fix issue where the ball could move out of the table and not return to
the centre. This could only occur if the horizontal speed and/or maximum
vertical speed were set to certain values, since the code previously
only checked whether the ball's x- and y-coordinates were equal to
certain minimum and maximum values, rather than less than (or equal to)
the minimum or greater than (or equal to) the maximum.